refactor: split game_list into widget and item components
